Adicionado tratamento global de exceções
