+++ src/main/java/com/codenation/controllers/UserController.java
@@ -1,7 +1,6 @@
 package com.codenation.controllers;
 
 import com.codenation.dtos.UserDTO;
-import com.codenation.exceptions.CustomizedExceptionHandlerResponse;
 import com.codenation.services.UserServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -15,8 +14,6 @@ import org.springframework.http.HttpStatus;
 import org.modelmapper.ModelMapper;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -43,11 +40,10 @@ public class UserController {
 
     @PostMapping
     @ApiOperation(value = "Cria um novo usuário")
-    public ResponseEntity<User> register(@RequestBody @Valid User user, Errors errors) {
+    public ResponseEntity<User> register(@RequestBody @Valid User user) {
         /*if (errors.hasErrors()) {
-            System.out.println(errors.getGlobalError().getDefaultMessage());
-            return ResponseEntity.badRequest().body(null);
-        }*/
+            //return ResponseEntity.badRequest().body(null);
+        } */
         User userCreated = userService.save(user);
         if (userCreated == null) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
@@ -61,16 +57,5 @@ public class UserController {
     public ResponseEntity<List<UserDTO>> getAll(){
         return ResponseEntity.status(HttpStatus.OK).body(userService.getAll()
             .stream().map(this::toUserDTO).collect(Collectors.toList()));
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<CustomizedExceptionHandlerResponse> handleException (MethodArgumentNotValidException exception) {
-        LocalDateTime time = LocalDateTime.now();
-        int code = HttpStatus.BAD_REQUEST.value();
-        String message = exception.getMessage();
-        String message2 = exception.getGlobalError().getDefaultMessage();
-        System.out.println(message2);
-        CustomizedExceptionHandlerResponse error = new CustomizedExceptionHandlerResponse(code, message, time);
-        return new ResponseEntity<CustomizedExceptionHandlerResponse>(error, HttpStatus.BAD_REQUEST);
     }
 }
